// types/types.ts
export type Categoty = {
    id: number;
    name: string;
    slug: string;
  };

  export type Post = {
    id: number;
    category: string;
    title: string;
    image: string;
    caption: string;
    date: string | Date;
    minutesToRead: number;
    author: string;
    nbViews: number;
    nbComments: number;
    slug: string;
    content?: string;
  };